+++ vector/src/main/java/im/vector/app/features/roomdirectory/createroom/CreateRoomController.kt
@@ -16,8 +16,10 @@ import im.vector.app.core.resources.StringProvider
 import im.vector.app.features.discovery.settingsSectionTitleItem
 import im.vector.app.features.form.formEditTextItem
 import im.vector.app.features.form.formEditableAvatarItem
+import im.vector.app.features.form.formEditableSquareAvatarItem
 import im.vector.app.features.form.formSubmitButtonItem
 import im.vector.app.features.form.formSwitchItem
+import im.vector.app.features.home.AvatarRenderer
 import im.vector.lib.strings.CommonStrings
 import org.matrix.android.sdk.api.MatrixConstants
 import org.matrix.android.sdk.api.extensions.orFalse
@@ -27,12 +29,11 @@ import javax.inject.Inject
 
 class CreateRoomController @Inject constructor(
         private val stringProvider: StringProvider,
+        private val avatarRenderer: AvatarRenderer,
         private val roomAliasErrorFormatter: RoomAliasErrorFormatter
 ) : TypedEpoxyController<CreateRoomViewState>() {
 
     var listener: Listener? = null
-
-    var index = 0
 
     override fun buildModels(viewState: CreateRoomViewState) {
         // display the form
@@ -41,17 +42,27 @@ class CreateRoomController @Inject constructor(
 
     private fun buildForm(viewState: CreateRoomViewState, enableFormElement: Boolean) {
         val host = this
-        formEditableAvatarItem {
-            id("avatar")
-            enabled(enableFormElement)
-            imageUri(viewState.avatarUri)
-            clickListener { host.listener?.onAvatarChange() }
-            deleteListener { host.listener?.onAvatarDelete() }
+        val isSpace = viewState.isSubSpace
+
+        if (isSpace) {
+            formEditableSquareAvatarItem {
+                id("avatar")
+                enabled(enableFormElement)
+                imageUri(viewState.avatarUri)
+                avatarRenderer(host.avatarRenderer)
+                clickListener { host.listener?.onAvatarChange() }
+                deleteListener { host.listener?.onAvatarDelete() }
+            }
+        } else {
+            formEditableAvatarItem {
+                id("avatar")
+                enabled(enableFormElement)
+                imageUri(viewState.avatarUri)
+                clickListener { host.listener?.onAvatarChange() }
+                deleteListener { host.listener?.onAvatarDelete() }
+            }
         }
-        settingsSectionTitleItem {
-            id("nameSection")
-            titleResId(CommonStrings.create_room_name_section)
-        }
+
         formEditTextItem {
             id("name")
             enabled(enableFormElement)
@@ -63,16 +74,16 @@ class CreateRoomController @Inject constructor(
                 host.listener?.onNameChange(text)
             }
         }
-        settingsSectionTitleItem {
-            id("topicSection")
-            titleResId(CommonStrings.create_room_topic_section)
-        }
         formEditTextItem {
             id("topic")
             enabled(enableFormElement)
             value(viewState.roomTopic)
             singleLine(false)
-            hint(host.stringProvider.getString(CommonStrings.create_room_topic_hint))
+            hint(
+                    host.stringProvider.getString(
+                            if (isSpace) CommonStrings.create_space_topic_hint else CommonStrings.create_room_topic_hint
+                    )
+            )
 
             onTextChange { text ->
                 host.listener?.onTopicChange(text)
@@ -81,7 +92,9 @@ class CreateRoomController @Inject constructor(
 
         settingsSectionTitleItem {
             id("visibility")
-            titleResId(CommonStrings.room_settings_room_access_title)
+            titleResId(
+                    if (isSpace) CommonStrings.room_settings_space_access_title else CommonStrings.room_settings_room_access_title
+            )
         }
 
         when (viewState.roomJoinRules) {
@@ -90,7 +103,23 @@ class CreateRoomController @Inject constructor(
                         id = "joinRule",
                         title = stringProvider.getString(CommonStrings.room_settings_room_access_private_title),
                         subtitle = stringProvider.getString(CommonStrings.room_settings_room_access_private_description),
-                        divider = false,
+                        divider = true,
+                        editable = true,
+                        action = { host.listener?.selectVisibility() }
+                )
+            }
+            RoomJoinRules.KNOCK -> {
+                buildProfileAction(
+                        id = "joinRule",
+                        title = stringProvider.getString(CommonStrings.room_settings_room_access_knock_title),
+                        subtitle = stringProvider.getString(
+                                if (isSpace) {
+                                    CommonStrings.room_settings_space_access_knock_description
+                                } else {
+                                    CommonStrings.room_settings_room_access_knock_description
+                                }
+                        ),
+                        divider = true,
                         editable = true,
                         action = { host.listener?.selectVisibility() }
                 )
@@ -99,8 +128,14 @@ class CreateRoomController @Inject constructor(
                 buildProfileAction(
                         id = "joinRule",
                         title = stringProvider.getString(CommonStrings.room_settings_room_access_public_title),
-                        subtitle = stringProvider.getString(CommonStrings.room_settings_room_access_public_description),
-                        divider = false,
+                        subtitle = stringProvider.getString(
+                                if (isSpace) {
+                                    CommonStrings.room_settings_space_access_public_description
+                                } else {
+                                    CommonStrings.room_settings_room_access_public_description
+                                }
+                        ),
+                        divider = true,
                         editable = true,
                         action = { host.listener?.selectVisibility() }
                 )
@@ -110,7 +145,7 @@ class CreateRoomController @Inject constructor(
                         id = "joinRule",
                         title = stringProvider.getString(CommonStrings.room_settings_room_access_restricted_title),
                         subtitle = stringProvider.getString(CommonStrings.room_create_member_of_space_name_can_join, viewState.parentSpaceSummary?.displayName),
-                        divider = false,
+                        divider = true,
                         editable = true,
                         action = { host.listener?.selectVisibility() }
                 )
@@ -122,7 +157,9 @@ class CreateRoomController @Inject constructor(
 
         settingsSectionTitleItem {
             id("settingsSection")
-            titleResId(CommonStrings.create_room_settings_section)
+            titleResId(
+                    if (isSpace) CommonStrings.create_space_settings_section else CommonStrings.create_room_settings_section
+            )
         }
 
         if (viewState.roomJoinRules == RoomJoinRules.PUBLIC) {
@@ -134,7 +171,11 @@ class CreateRoomController @Inject constructor(
                 suffixText(":" + viewState.homeServerName)
                 prefixText("#")
                 maxLength(MatrixConstants.maxAliasLocalPartLength(viewState.homeServerName))
-                hint(host.stringProvider.getString(CommonStrings.room_alias_address_hint))
+                hint(
+                        host.stringProvider.getString(
+                                if (isSpace) CommonStrings.create_space_alias_hint else CommonStrings.room_alias_address_hint
+                        )
+                )
                 errorMessage(
                         host.roomAliasErrorFormatter.format(
                                 (((viewState.asyncCreateRoomRequest as? Fail)?.error) as? CreateRoomFailure.AliasError)?.aliasError
@@ -144,10 +185,10 @@ class CreateRoomController @Inject constructor(
                     host.listener?.setAliasLocalPart(value)
                 }
             }
-        } else {
             dividerItem {
-                id("divider0")
+                id("dividerAlias")
             }
+        } else {
             // Room encryption for private room
             formSwitchItem {
                 id("encryption")
@@ -156,6 +197,8 @@ class CreateRoomController @Inject constructor(
                 summary(
                         if (viewState.hsAdminHasDisabledE2E) {
                             host.stringProvider.getString(CommonStrings.settings_hs_admin_e2e_disabled)
+                        } else if (isSpace) {
+                            host.stringProvider.getString(CommonStrings.create_space_encryption_description)
                         } else {
                             host.stringProvider.getString(CommonStrings.create_room_encryption_description)
                         }
@@ -167,19 +210,48 @@ class CreateRoomController @Inject constructor(
                     host.listener?.setIsEncrypted(value)
                 }
             }
+
+            dividerItem {
+                id("dividerEncryption")
+            }
+
+            formSwitchItem {
+                id("encryptedState")
+                enabled(enableFormElement && viewState.encryptionEnabled)
+                title(host.stringProvider.getString(CommonStrings.create_room_encrypted_state_title))
+                summary(host.stringProvider.getString(CommonStrings.create_room_encrypted_state_warning))
+                switchChecked(viewState.encryptStateEvents && viewState.encryptionEnabled)
+                listener { value -> host.listener?.setEncryptStateEvents(value) }
+            }
+
+            dividerItem {
+                id("dividerEncryptedState")
+            }
+        }
+
+        formSwitchItem {
+            id("federation")
+            enabled(enableFormElement)
+            title(host.stringProvider.getString(CommonStrings.create_room_disable_federation_title, viewState.homeServerName))
+            summary(host.stringProvider.getString(CommonStrings.create_room_disable_federation_description))
+            switchChecked(viewState.disableFederation)
+            listener { value -> host.listener?.setDisableFederation(value) }
+        }
+
+        dividerItem {
+            id("dividerFederation")
         }
 
         buildAdvancedRoomOptions(
                 options = viewState,
                 stringProvider = stringProvider,
-                homeServerName = viewState.homeServerName,
                 enabled = enableFormElement,
                 listener = listener,
         )
 
         formSubmitButtonItem {
             id("submit")
-            enabled(enableFormElement)
+            enabled(enableFormElement && (!isSpace || viewState.roomName.isNotBlank()))
             buttonTitleId(CommonStrings.create_room_action_create)
             buttonClickListener { host.listener?.submit() }
         }
@@ -193,6 +265,8 @@ class CreateRoomController @Inject constructor(
         fun selectVisibility()
         fun setAliasLocalPart(aliasLocalPart: String)
         fun setIsEncrypted(isEncrypted: Boolean)
+        fun setEncryptStateEvents(enabled: Boolean)
+        fun setDisableFederation(disableFederation: Boolean)
         fun submit()
     }
 }